% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_dates}
\alias{fetch_dates}
\title{fetch_dates}
\usage{
fetch_dates(start_date, end_date, output_dir, replace = FALSE,
  quiet = FALSE)
}
\arguments{
\item{start_date}{The string (yyyy-mm-dd) for start of queried range}

\item{end_date}{The string (yyyy-mm-dd) for end of queried range}

\item{output_dir}{An existing directory to store downloaded files}

\item{replace}{Boolean to overwrite existing files}

\item{quiet}{Boolean to omit download messages}
}
\value{
invisible(output_paths) List of file paths to downloaded files
}
\description{
Fetch data from Valley Bike service for specified dates.
}
\details{
Fetch data from Valley Bike service for specified dates. Provide a
start date and an endate as well as a place to save the files that
will be downloaded. The list of file paths for the downloaded
files are returned invisibly.
}
