% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes}
\alias{routes}
\title{Routes}
\format{A data frame with 189,419 rows and 6 variables:
\describe{
  \item{route_id}{identifier for specific route}
  \item{bike}{unique bike number}
  \item{date}{date and time of entry}
  \item{latitude}{latitude coordinate of entry}
  \item{longitude}{longitude coordinate of entry}
  \item{user_id}{identifier for specific user}
  ...
}}
\usage{
routes
}
\description{
A dataset containing the first and last entries of all bike rides.
}
\keyword{datasets}
