% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations}
\format{A data frame with 54 rows and 6 variables:
\describe{
  \item{serial_num}{identifier for station}
  \item{address}{station address}
  \item{station_name}{station name}
  \item{num_docks}{the number of docks in the station}
  \item{latitude}{latitude coordinate of station}
  \item{longitude}{longitude coordinate of station}
  \item{community_name}{name of Pioneer Valley community}
}}
\usage{
stations
}
\description{
A dataset containing the Valley Bike stations.
}
\keyword{datasets}
