% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleybike.R
\docType{package}
\name{valleybike}
\alias{valleybike}
\title{valleybike: A package for exploring data from Pioneer Valley bikeshare.}
\description{
This package provides an easy way to start exploring data from
the Valley Bike initiative through the `stations` and `routes`
datasets.
}
