% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValleyBikes.R
\docType{package}
\name{ValleyBikes}
\alias{ValleyBikes}
\alias{ValleyBikes-package}
\title{ValleyBikes: A package for exploring data from Pioneer Valley bikeshare.}
\description{
The foo package provides an easy way to start exploring data from
the Valley Bike initiative through the `stations` and `routes`
datasets.
}
