% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations}
\format{A data frame with 54 rows and 6 variables:
\describe{
  \item{Serial_Num}{identifier for station}
  \item{Address}{station address}
  \item{Station_Name}{station name}
  \item{Num_Docks}{the number of docks in the station}
  \item{Latitude}{latitude coordinate of station}
  \item{Longitude}{longitude coordinate of station}
  
}}
\usage{
stations
}
\description{
A dataset containing the Valley Bike stations.
}
\keyword{datasets}
